update bao sach da mat

See merge request group6-se1805-swp391/SWP391_Library!27

+++ Library/src/main/java/com/springdemo/library/model/dto/MissingSach.java
@@ -8,5 +8,5 @@ import lombok.Getter;
 @Getter
 public class MissingSach {
     private Sach missingSach;
-    private int soLuongDaMat;
+    private long soLuongDaMat;
 }
